+++ test/RetroPunks.t.sol
@@ -4,11 +4,11 @@ pragma solidity ^0.8.30;
 import "forge-std/Test.sol";
 import "../src/RetroPunks.sol";
 import "../src/interfaces/ISVGRenderer.sol";
-import { IERC721Receiver } from "@openzeppelin/contracts/token/ERC721/IERC721Receiver.sol";
+import {IERC721Receiver} from "@openzeppelin/contracts/token/ERC721/IERC721Receiver.sol";
 
 // --- Mock Renderer ---
 contract MockRenderer is ISVGRenderer {
-    function renderSVG(uint16, uint8, uint) external pure override returns (string memory, string memory) {
+    function renderSVG(uint16, uint8, uint256) external pure override returns (string memory, string memory) {
         return ("<svg>mock</svg>", '"attributes": []');
     }
 }
@@ -17,7 +17,7 @@ contract MockRenderer is ISVGRenderer {
 contract RetroPunksTest is Test, IERC721Receiver {
     RetroPunks public retroPunks;
     MockRenderer public renderer;
-    
+
     address public owner;
     address public user1;
     address[] public allowedSeaDrops;
@@ -50,13 +50,7 @@ contract RetroPunksTest is Test, IERC721Receiver {
 
         renderer = new MockRenderer();
 
-        retroPunks = new RetroPunks(
-            ISVGRenderer(address(renderer)),
-            GLOBAL_SEED_HASH,
-            SHUFFLER_SEED_HASH,
-            MAX_SUPPLY,
-            allowedSeaDrops
-        );
+        retroPunks = new RetroPunks(ISVGRenderer(address(renderer)), GLOBAL_SEED_HASH, SHUFFLER_SEED_HASH, MAX_SUPPLY, allowedSeaDrops);
     }
 
     // ==========================================
@@ -80,14 +74,14 @@ contract RetroPunksTest is Test, IERC721Receiver {
         emit ShufflerSeedRevealed(SHUFFLER_SEED);
 
         retroPunks.revealShufflerSeed(SHUFFLER_SEED, SHUFFLER_NONCE);
-        
+
         assertTrue(retroPunks.shufflerSeedRevealed());
         assertEq(retroPunks.shufflerSeed(), SHUFFLER_SEED);
     }
 
     function testRevealShufflerSeedInvalid() public {
         vm.expectRevert(RetroPunks.InvalidShufflerSeedReveal.selector);
-        retroPunks.revealShufflerSeed(SHUFFLER_SEED, 0); 
+        retroPunks.revealShufflerSeed(SHUFFLER_SEED, 0);
     }
 
     function testRevealGlobalSeed() public {
@@ -95,7 +89,7 @@ contract RetroPunksTest is Test, IERC721Receiver {
         emit GlobalSeedRevealed(GLOBAL_SEED);
 
         retroPunks.revealGlobalSeed(GLOBAL_SEED, GLOBAL_NONCE);
-        
+
         assertTrue(retroPunks.globalSeedRevealed());
         assertEq(retroPunks.globalSeed(), GLOBAL_SEED);
     }
@@ -103,7 +97,7 @@ contract RetroPunksTest is Test, IERC721Receiver {
     function testCloseMint() public {
         retroPunks.closeMint();
         assertTrue(retroPunks.mintIsClosed());
-        
+
         retroPunks.revealShufflerSeed(SHUFFLER_SEED, SHUFFLER_NONCE);
         vm.expectRevert(RetroPunks.MintIsClosed.selector);
         retroPunks.ownerMint(address(this), 1);
@@ -116,9 +110,9 @@ contract RetroPunksTest is Test, IERC721Receiver {
     function testOwnerMint() public {
         retroPunks.revealShufflerSeed(SHUFFLER_SEED, SHUFFLER_NONCE);
         retroPunks.ownerMint(address(this), 5);
-        
+
         assertEq(retroPunks.balanceOf(address(this)), 5);
-        assertEq(retroPunks.ownerMintsRemaining(), 5); 
+        assertEq(retroPunks.ownerMintsRemaining(), 5);
         assertEq(retroPunks.totalSupply(), 5);
     }
 
@@ -142,17 +136,17 @@ contract RetroPunksTest is Test, IERC721Receiver {
         retroPunks.ownerMint(address(this), 1);
         uint256 tokenId = 1;
 
-        (uint16 seed, , , ) = retroPunks.globalTokenMetadata(tokenId);
-        if (seed < 7) return; 
+        (uint16 seed,,,) = retroPunks.globalTokenMetadata(tokenId);
+        if (seed < 7) return;
 
         string memory newName = "CyberPunk 2077";
-        
+
         vm.expectEmit(true, false, false, true);
         emit NameChanged(tokenId, newName, address(this));
-        
+
         retroPunks.setTokenName(tokenId, newName);
-        
-        (, , string memory storedName, ) = retroPunks.globalTokenMetadata(tokenId);
+
+        (,, string memory storedName,) = retroPunks.globalTokenMetadata(tokenId);
         assertEq(storedName, newName);
     }
 
@@ -161,17 +155,17 @@ contract RetroPunksTest is Test, IERC721Receiver {
         retroPunks.ownerMint(address(this), 1);
         uint256 tokenId = 1;
 
-        (uint16 seed, , , ) = retroPunks.globalTokenMetadata(tokenId);
+        (uint16 seed,,,) = retroPunks.globalTokenMetadata(tokenId);
         if (seed < 7) return;
 
         string memory newBio = "Living on the blockchain";
-        
+
         vm.expectEmit(true, false, false, true);
         emit BioChanged(tokenId, newBio, address(this));
-        
+
         retroPunks.setTokenBio(tokenId, newBio);
-        
-        (, , , string memory storedBio) = retroPunks.globalTokenMetadata(tokenId);
+
+        (,,, string memory storedBio) = retroPunks.globalTokenMetadata(tokenId);
         assertEq(storedBio, newBio);
     }
 
@@ -180,14 +174,14 @@ contract RetroPunksTest is Test, IERC721Receiver {
         retroPunks.ownerMint(address(this), 1);
         uint256 tokenId = 1;
 
-        (uint16 seed, , , ) = retroPunks.globalTokenMetadata(tokenId);
+        (uint16 seed,,,) = retroPunks.globalTokenMetadata(tokenId);
         if (seed < 7) return;
 
         string memory longName = "This name is definitely way too long for the limit";
         vm.expectRevert(RetroPunks.NameIsTooLong.selector);
         retroPunks.setTokenName(tokenId, longName);
 
-        string memory invalidName = "Punk@Home"; 
+        string memory invalidName = "Punk@Home";
         vm.expectRevert(RetroPunks.InvalidCharacterInName.selector);
         retroPunks.setTokenName(tokenId, invalidName);
     }
@@ -197,17 +191,17 @@ contract RetroPunksTest is Test, IERC721Receiver {
         retroPunks.ownerMint(address(this), 1);
         uint256 tokenId = 1;
 
-        (uint16 seed, , , ) = retroPunks.globalTokenMetadata(tokenId);
+        (uint16 seed,,,) = retroPunks.globalTokenMetadata(tokenId);
         if (seed < 7) return;
 
-        uint8 newBgIndex = 1; 
-        
+        uint8 newBgIndex = 1;
+
         vm.expectEmit(true, false, false, false);
         emit MetadataUpdate(tokenId);
 
         retroPunks.setTokenBackground(tokenId, newBgIndex);
 
-        (, uint8 bgIndex, , ) = retroPunks.globalTokenMetadata(tokenId);
+        (, uint8 bgIndex,,) = retroPunks.globalTokenMetadata(tokenId);
         assertEq(bgIndex, newBgIndex);
     }
 
@@ -216,10 +210,10 @@ contract RetroPunksTest is Test, IERC721Receiver {
         retroPunks.ownerMint(address(this), 1);
         uint256 tokenId = 1;
 
-        (uint16 seed, , , ) = retroPunks.globalTokenMetadata(tokenId);
+        (uint16 seed,,,) = retroPunks.globalTokenMetadata(tokenId);
         if (seed < 7) return;
 
-        vm.prank(user1); 
+        vm.prank(user1);
         vm.expectRevert(RetroPunks.CallerIsNotTokenOwner.selector);
         retroPunks.setTokenName(tokenId, "Hacked Name");
     }
@@ -231,7 +225,7 @@ contract RetroPunksTest is Test, IERC721Receiver {
     function testTokenURI() public {
         retroPunks.revealShufflerSeed(SHUFFLER_SEED, SHUFFLER_NONCE);
         retroPunks.ownerMint(address(this), 1);
-        
+
         string memory uri = retroPunks.tokenURI(1);
         assertTrue(bytes(uri).length > 0);
     }
@@ -241,9 +235,9 @@ contract RetroPunksTest is Test, IERC721Receiver {
         retroPunks.ownerMint(address(this), 1);
 
         (uint16 tokenIdSeed, uint8 backgroundIndex, string memory name, string memory bio) = retroPunks.globalTokenMetadata(1);
-        
-        assertTrue(tokenIdSeed < MAX_SUPPLY); 
-        assertEq(backgroundIndex, retroPunks.defaultBackgroundIndex()); 
+
+        assertTrue(tokenIdSeed < MAX_SUPPLY);
+        assertEq(backgroundIndex, retroPunks.defaultBackgroundIndex());
         assertTrue(bytes(name).length > 0);
         assertEq(bio, "A RetroPunk living on-chain.");
     }
